requirements.txt
glfw==2.1.0
requests==2.23.0
scipy==1.3.1
tqdm==4.45.0
opencv_python==4.2.0.34
torch==1.6.0+cpu
torchvision==0.7.0+cpu
pyrr==0.10.3
numpy==1.18.2
Django==3.1.4
scikit_image==0.16.2
ipython==7.22.0
Pillow==8.2.0
pyexr==0.3.9
PyOpenGL==3.1.5
pytorch3d==0.4.0
skimage==0.0
trimesh==3.9.14
